+++ Sprocket/CMakeLists.txt
@@ -43,7 +43,7 @@ add_library(Sprocket STATIC
             UI/SimpleUI.cpp
             UI/SinglePanelUI.cpp
             UI/DevUI.cpp
-            UI/ImGuiHelpers.cpp
+            UI/ImGuiXtra.cpp
 
             Scene/Entity.cpp
             Scene/Scene.cpp
